Fantasy本次优化调整如下
1. 优化Scene的ThreadPool模式运行，改为FIFO机制能提供更好的公平性。
2. CoroutineLockComponent.Release的锁类型参数从int变成long。
3. Scene.GetSession从Public改为internal，因为这个方法只是框架内使用的。
4. KCP移除mimalloc相关的逻辑,因为很多人都会提示缺少库的问题。
5. Serialize增加两个返回byte的接口。
6. Directory.GetCurrentDirectory()替换为AppContext.BaseDirectory。
  因为Directory.GetCurrentDirectory() 返回的是当前工作目录（Current Working Directory），而不是程序的可执行文件所在目录。
7. 启动框架的时候会打印一个当前使用版本号。
8. 服务器在心跳超时后默认不在打印超时的警告日志。
9. 优化了网络消息处理的Unsafe代码。
10. 移除Fantasy.Console相关支持，目前精力不够，后续有机会再增加。

+++ Fantasy.Unity/Fantasy.Unity/Runtime/Core/Entitas/Component/CoroutineLock/CoroutineLockComponent.cs
@@ -82,7 +82,7 @@ namespace Fantasy.Async
         /// </summary>
         /// <param name="coroutineLockType"></param>
         /// <param name="coroutineLockQueueKey"></param>
-        public void Release(int coroutineLockType, long coroutineLockQueueKey)
+        public void Release(long coroutineLockType, long coroutineLockQueueKey)
         {
             if (IsDisposed)
             {
